[REFACTOR]: 중복 코드 공통 상위 컴포넌트로 이동
